+++ src/mailman/handlers/tests/test_recipients.py
@@ -198,3 +198,24 @@ To: test-owner@example.com
         msgdata = {}
         self._process(self._mlist, self._msg, msgdata)
         self.assertEqual(msgdata['recipients'], set(('noreply@example.com',)))
+
+    def test_site_admin_unicode(self):
+        # Since the config file is read as bytes, the site_owner is also a
+        # bytes and must be converted to unicode when used as a fallback.
+        self._cris.unsubscribe()
+        self._dave.unsubscribe()
+        self.assertEqual(self._mlist.administrators.member_count, 0)
+        msgdata = {}
+        # In order to properly mimic the testing environment, use
+        # config.push()/config.pop() directly instead of using the
+        # configuration() context manager.
+        config.push('test_site_admin_unicode', b"""\
+[mailman]
+site_owner: siteadmin@example.com
+""")
+        try:
+            self._process(self._mlist, self._msg, msgdata)
+        finally:
+            config.pop('test_site_admin_unicode')
+        self.assertEqual(len(msgdata['recipients']), 1)
+        self.assertIsInstance(list(msgdata['recipients'])[0], unicode)