+++ src/mailman/docs/NEWS.rst
@@ -12,6 +12,14 @@ Here is a history of user visible changes to Mailman.
 ====================================
 (2014-XX-XX)
 
+Bugs
+----
+ * Fixed Unicode errors in the digest runner and when sending messages to the
+   site owner as a fallback.  Given by Aurélien Bompard.  (LP: #1130957).
+ * Fix Unicode errors when a message being added to the digest has non-ascii
+   characters in its payload, but no Content-Type header defining a charset.
+   Given by Aurélien Bompard.  (LP: #1170347)
+
 Commands
 --------
  * The `mailman conf` command no longer takes the `-t/--sort` option; the